- Remove unnecessary loading and dotenv import - Move test_connection to DBConnector
