+++ margin_app/app/crud/base.py
@@ -9,6 +9,7 @@ from contextlib import asynccontextmanager
 from typing import AsyncIterator, Callable, Type, TypeVar, List, Optional
 
 from sqlalchemy import select
+from sqlalchemy.sql import text
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
@@ -160,4 +161,14 @@ class DBConnector:
 
             result = await db.execute(query)
             return result.scalars().all()
-          
+    
+
+    async def test_connection(self):
+        """
+        Test the database connection.
+        :return
+        """
+        async with self.session() as session:
+            result = await session.execute(text("SELECT version()"))
+            return f"PostgreSQL version: {result.scalar()}"
+
